[DEV-56] feat: grid로 dashboard 페이지 크기 지정
