+++ src/app/dashboard/page.tsx
@@ -1,6 +1,8 @@
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 
-import RecentlyTodoContainer from "@/containers/dashboard/recently-todo-container";
+import MyProgressContainer from "@/containers/dashboard/my-progress/my-progress-container";
+import RecentlyTodoContainer from "@/containers/dashboard/recently-todo/recently-todo-container";
+import TodosByGoalContainer from "@/containers/dashboard/todos-by-goal/todos-by-goal-container";
 import getQueryClient from "@/lib/get-query-client";
 import { getDashboardOptions } from "@/services/dashboard";
 
@@ -10,11 +12,15 @@ export default async function DashboardPage() {
   queryClient.prefetchQuery(getDashboardOptions());
 
   return (
-    <main className="mx-auto min-h-screen w-full bg-slate-100 px-16">
-      <h1 className="hidden font-semibold md:block">대시보드</h1>
-      <HydrationBoundary state={dehydrate(queryClient)}>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <main className="grid h-full min-h-screen w-full grid-rows-[repeat(2,_258px)_1fr] gap-x-24 gap-y-8 overflow-y-scroll bg-slate-100 px-16 md:grid-cols-2 md:grid-rows-[auto_250px_1fr]">
+        <h1 className="hidden font-semibold md:col-span-2 md:block">
+          대시보드
+        </h1>
         <RecentlyTodoContainer />
-      </HydrationBoundary>
-    </main>
+        <MyProgressContainer />
+        <TodosByGoalContainer />
+      </main>
+    </HydrationBoundary>
   );
 }
