+++ src/containers/dashboard/my-progress/my-progress-container.tsx
@@ -1,0 +1,5 @@
+export default function MyProgressContainer() {
+  return (
+    <section className="relative flex h-258 w-full flex-col gap-16 rounded-2xl bg-white p-16 md:h-250" />
+  );
+}
